+++ R/mm_is_valid_day.R
@@ -14,10 +14,14 @@
 #'   worth of data is valid for modeling. \code{full_day}: Do the data span the 
 #'   full expected period (e.g., from 10:30pm on preceding day to 6am on 
 #'   following day)? \code{even_timesteps}: are all of the timesteps within the 
-#'   day the same length, to within a tolerance of 0.2% of the timestep length? 
+#'   day the same length, to within a tolerance of 0.2\% of the timestep length? 
 #'   \code{complete_data}: are all columns of input data available at every 
-#'   timestep? \code{pos_discharge}: is discharge greater than 0 at every
-#'   timestep?
+#'   timestep? \code{pos_discharge}: is discharge greater than 0 at every 
+#'   timestep? A further test is implied if \code{required_timestep} is a non-NA
+#'   numeric.
+#' @param required_timestep NA or numeric (length 1). If numeric, the timestep 
+#'   length in days that a date must have to pass the validity check (to within 
+#'   a tolerance of 0.2\% of the value of \code{required_timestep})
 #' @param ply_date the Date this data_ply is intended to match. May be NA
 #' @param timestep_days the expected timestep length in fraction of a day; for 
 #'   example, a 1-hour timestep is 1/24 is 0.0416667. This is calculated within 
@@ -37,6 +41,7 @@ mm_is_valid_day <- function(
   data_ply, # inheritParams mm_model_by_ply_prototype
   day_start=4, day_end=27.99, # inheritParams mm_model_by_ply
   day_tests=c('full_day', 'even_timesteps', 'complete_data', 'pos_discharge'), 
+  required_timestep=NA,
   ply_date=as.Date(format(data_ply[nrow(data_ply)/2,'solar.time'], "%Y-%m-%d")),
   timestep_days=NA
 ) {
@@ -51,7 +56,7 @@ mm_is_valid_day <- function(
   stop_strs <- character(0)
   
   # find the mean timestep if it will be needed
-  if(any(c('full_day','even_timesteps') %in% day_tests)) {
+  if(any(c('full_day','even_timesteps') %in% day_tests) || !isTRUE(is.na(required_timestep))) {
     timestep.days <- if(!is.null(timestep_days) && is.na(timestep_days)) {
       tryCatch(
         mm_get_timestep(data_ply$solar.time, format='mean'),
@@ -84,8 +89,8 @@ mm_is_valid_day <- function(
   }
   
   # Require that on each day solar.time has a ~single, ~consistent time step
+  # (with tolerance of 0.2% of mean timestep length)
   if('even_timesteps' %in% day_tests & is.finite(timestep.days)) {
-    # require a single unique timestep with tolerance of 2% of mean timestep length
     tryCatch(
       ts <- mm_get_timestep(data_ply$solar.time, format='unique', require_unique=TRUE, tol=0.002*timestep.days),
       error=function(e) {
@@ -93,16 +98,27 @@ mm_is_valid_day <- function(
       })
   }
   
+  # Require the actual average timestep to match the value specified in
+  # required_timestep (within tolerance of 0.2% of required_timestep length)
+  if(!isTRUE(is.na(required_timestep)) & is.finite(timestep.days)) {
+    if(length(required_timestep) != 1 || !is.numeric(required_timestep)) {
+      stop('expecting required_timestep as single numeric value (or NA)')
+    }
+    if(!isTRUE(all.equal(timestep.days - required_timestep, tolerance=0.002*required_timestep))) {
+      stop_strs <- c(stop_strs, sprintf("timestep != %0.08f d", required_timestep))
+    }
+  }
+  
+  # Require complete data in all columns
   if('complete_data' %in% day_tests) {
-    # Require complete data in all columns
     for(col in names(data_ply)) {
       if(any(is.na(data_ply[[col]]))) 
         stop_strs <- c(stop_strs, paste0("NAs in ", col))
     }
   }
   
+  # Require discharge (if present) to be positive at all times
   if('pos_discharge' %in% day_tests) {
-    # require discharge (if present) to be positive at all times
     if('discharge' %in% names(data_ply) && any(!is.na(data_ply$discharge))) {
       if(any(data_ply$discharge[which(!is.na(data_ply$discharge))] <= 0)) 
         stop_strs <- c(stop_strs, "discharge <= 0")