+++ NEWS.md
@@ -1,6 +1,19 @@
+# 0.9.44
+
+* New specs element `required_timestep` allows you to require that each date has
+the specified numeric timestep in days
+
+# 0.9.42-0.9.43
+
+* Bugfixes for new test that excludes days with non-positive discharge
+
+# 0.9.40
+
+* Update to roxygen2 6.0.1
+
 # 0.9.39
 
-* Bayesian, MLE, and nighttime regression models can now all test for and
+* Bayesian, MLE, and nighttime regression models can now all test for and 
 exclude days with non-positive discharge
 
 # 0.9.36